feat: Add OTPGenerator class for generating One-Time Passwords
This commit adds the `OTPGenerator` class to the `msft_aama` module. The `OTPGenerator` class is responsible for generating One-Time Passwords (OTP) using a secret key. It provides a method `generate_otp` to generate a new OTP. This addition enhances the functionality of the `msft_aama` module by allowing the generation of OTPs for authentication purposes.