chore: Update dependencies in pyproject.toml and pdm.lock
This commit updates the `pyproject.toml` and `pdm.lock` files to include the latest versions of the dependencies. Specifically, the `jsonschema` package has been updated to version `4.22.0`, and the `requests-mock` package has been updated to version `1.12.1`. This update ensures that the project uses the most recent versions of these packages, providing improved functionality and compatibility.