refactor: Add additional verification method to PostRequest class
This commit adds an additional verification method, `_additional_verify`, to the `PostRequest` class in the `msft_aama` module. This method is used for additional verification of the response data. It takes a `response` dictionary as an argument and performs the necessary checks. This enhancement improves the reliability and accuracy of the response data validation process in the `PostRequest` class.