feat: Add cookies parameter to AddSecurityInfo constructor
This commit adds a new optional `cookies` parameter to the `AddSecurityInfo` class constructor in the `msft_aama` module. The `cookies` parameter allows for passing a dictionary of cookies to be included in the request. This enhancement provides flexibility in customizing the request headers and improves the functionality of the `AddSecurityInfo` class.