feat: Register new security method for the user
This commit adds a new function `register` to the `__init__.py` file in the `msft_aama` module. The `register` function is responsible for registering a new security method for the user. It performs the following steps:
1. Initializes the mobile app registration.
2. Adds security info.
3. Verifies the security info.

This enhancement improves the functionality of the `msft_aama` module by providing a convenient way to register a new security method for the user.