feat: Update requests library version and add requests-oauthlib
The code changes update the `pdm.lock` and `pyproject.toml` files to include the latest version of the `requests` library and add the `requests-oauthlib` library as a dependency. This update improves the functionality of the project by ensuring compatibility with the latest version of the `requests` library and adding support for OAuth authentication.