refactor: Add flake8 noqa comment to ignore F401 error
The code changes add a flake8 noqa comment to ignore the F401 error in the `__init__.py` file. This error occurs when an imported module or variable is not used in the code. The comment is added to indicate that the error is intentionally ignored in this case. This improves the code quality and readability.